Add graphsage withoud self-loop

+++ paddle/phi/kernels/gpu/graph_reindex_funcs.h
@@ -24,7 +24,7 @@ namespace phi {
 
 template <typename T>
 inline __device__ size_t Hash(T id, int64_t size) {
-  return id % size;
+  return static_cast<unsigned long long int>(id) % size;
 }
 
 template <typename T>
